// data/github/current/ui/packages/billing-app/components/pickers/__generated__/PaginatedRepositoryListQuery.graphql.ts
export type PaginatedRepositoryListQuery$variables = {
  after?: string | null | undefined;
  excludeArchived?: boolean | null | undefined;
  first?: number | null | undefined;
  id: string;
  phrase?: string | null | undefined;
  slug: string;
};
export type PaginatedRepositoryListQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"PaginatedRepositoryPickerRepos">;
  } | null | undefined;
};
export type PaginatedRepositoryListQuery = {
  response: PaginatedRepositoryListQuery$data;
  variables: PaginatedRepositoryListQuery$variables;
};